Fix tray icon menu position and color
