+++ GUI/Main.cs
@@ -192,6 +192,7 @@ namespace CKAN
             settingsToolStripMenuItem.DropDown.Renderer = new FlatToolStripRenderer();
             helpToolStripMenuItem.DropDown.Renderer = new FlatToolStripRenderer();
             FilterToolButton.DropDown.Renderer = new FlatToolStripRenderer();
+            this.minimizedContextMenuStrip.Renderer = new FlatToolStripRenderer();
 
             // We need to initialize the error dialog first to display errors.
             errorDialog = controlFactory.CreateControl<ErrorDialog>();
@@ -1130,113 +1131,6 @@ namespace CKAN
         {
             UpdateTrayState();
         }
-
-        private void minimizeNotifyIcon_MouseDoubleClick(object sender, MouseEventArgs e)
-        {
-            OpenWindow();
-        }
-
-        private void updatesToolStripMenuItem_Click(object sender, EventArgs e)
-        {
-            OpenWindow();
-            MarkAllUpdatesToolButton_Click(sender, e);
-        }
-
-        private void refreshToolStripMenuItem_Click(object sender, EventArgs e)
-        {
-            UpdateRepo();
-        }
-
-        private void pauseToolStripMenuItem_Click(object sender, EventArgs e)
-        {
-            configuration.RefreshPaused = !configuration.RefreshPaused;
-            if (configuration.RefreshPaused)
-            {
-                refreshTimer.Stop();
-                pauseToolStripMenuItem.Text = "Resume";
-            }
-            else
-            {
-                refreshTimer.Start();
-                pauseToolStripMenuItem.Text = "Pause";
-            }
-        }
-
-        private void openCKANToolStripMenuItem_Click(object sender, EventArgs e)
-        {
-            OpenWindow();
-        }
-
-        private void cKANSettingsToolStripMenuItem1_Click(object sender, EventArgs e)
-        {
-            OpenWindow();
-            new SettingsDialog().ShowDialog();
-        }
-
-        #region Tray Behaviour
-
-        public void CheckTrayState()
-        {
-            enableTrayIcon = configuration.EnableTrayIcon;
-            minimizeToTray = configuration.MinimizeToTray;
-            pauseToolStripMenuItem.Enabled = winReg.RefreshRate != 0;
-            pauseToolStripMenuItem.Text = configuration.RefreshPaused ? "Resume" : "Pause";
-            UpdateTrayState();
-        }
-
-        private void UpdateTrayState()
-        {
-            if (enableTrayIcon)
-            {
-                minimizeNotifyIcon.Visible = true;
-
-                if (WindowState == FormWindowState.Minimized)
-                {
-                    if (minimizeToTray)
-                    {
-                        // Remove our taskbar entry
-                        Hide();
-                    }
-                }
-                else
-                {
-                    // Save the window state
-                    configuration.IsWindowMaximised = WindowState == FormWindowState.Maximized;
-                    configuration.Save();
-                }
-            }
-            else
-            {
-                minimizeNotifyIcon.Visible = false;
-            }
-        }
-
-        public void UpdateTrayInfo()
-        {
-            var count = mainModList.CountModsByFilter(GUIModFilter.InstalledUpdateAvailable);
-
-            if (count == 0)
-            {
-                updatesToolStripMenuItem.Enabled = false;
-                updatesToolStripMenuItem.Text = "No available updates";
-            }
-            else
-            {
-                updatesToolStripMenuItem.Enabled = true;
-                updatesToolStripMenuItem.Text = $"{count} available update" + (count == 1 ? "" : "s");
-            }
-        }
-
-        /// <summary>
-        /// Open the GUI and set it to the correct state.
-        /// </summary>
-        public void OpenWindow()
-        {
-            Show();
-            WindowState = configuration.IsWindowMaximised ? FormWindowState.Maximized : FormWindowState.Normal;
-        }
-
-        #endregion
 
         #region Navigation History
